🎨 (app) Updated Notification Structure
