+++ libs/app/notifications/utils/README.md
@@ -1,0 +1,7 @@
+# app-notifications-utils
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test app-notifications-utils` to execute the unit tests.
