+++ libs/app/notifications/utils/src/interfaces/notifications.interface.ts
@@ -1,0 +1,6 @@
+export interface INotification {
+    userName: string;
+    profilePictureUrl: string;
+    comment: string;
+    recipeId: string;
+  }
