+++ tsconfig.base.json
@@ -57,6 +57,9 @@
       "@fridge-to-plate/app/notifications/ui": [
         "libs/app/notifications/ui/src/index.ts"
       ],
+      "@fridge-to-plate/app/notifications/utils": [
+        "libs/app/notifications/utils/src/index.ts"
+      ],
       "@fridge-to-plate/app/preferences/data-access": [
         "libs/app/preferences/data-access/src/index.ts"
       ],
